+++ src/main/java/fr/tropweb/miningmanager/exception/EconomyException.java
@@ -1,4 +1,0 @@
-package fr.tropweb.miningmanager.exception;
-
-public class EconomyException extends RuntimeException {
-}